Fix werror: pointless comparison of unsigned integer with zero

+++ core/unit_test/TestExecutionSpace.hpp
@@ -56,8 +56,8 @@ struct CheckClassWithExecutionSpaceAsDataMemberIsCopyable {
   KOKKOS_FUNCTION void operator()(int, int& e) const {
     auto copy = *this;
     // not actually doing anything useful, mostly checking that
-    // ExecutionSpace::in_parallel() is callalable
-    if (copy.device.in_parallel() < 0) {
+    // ExecutionSpace::in_parallel() is callable
+    if (static_cast<int>(copy.device.in_parallel()) < 0) {
       ++e;
     }
   }
